fix(swscale): fix Windows-only SWS_* type mismatch and add modern sws_scale_frame API
FFmpeg 8 turned the SWS_* constants from #defines into an enum
(SwsFlags), so bindgen now generates them with the target's enum ABI:
`i32` on MSVC but `u32` elsewhere. This made every `flags: u32`
argument of SwsContext::get_context()/get_cached_context() fail to
compile on the Windows CI jobs. Fix all call sites with a plain
`as u32` cast, which is valid on both ABIs.

Additionally, expose the modern libswscale API as new methods, without
touching any existing ones:

- SwsContext::alloc() (sws_alloc_context)
- SwsContext::scale_full_frame(dst, src) (sws_scale_frame)

Both are gated behind the new `ffmpeg8` feature rather than `ffmpeg7`:
although sws_scale_frame() was introduced in FFmpeg 7, it segfaults on
all 7.x releases due to an upstream buffer pool bug (fixed in FFmpeg 8
by 6b402cdbf46e4398b3285277f3ff7c3654d57ce6).

The thumbnail test now uses the modern API on FFmpeg 8+ and falls back
to the classic sws_getContext()/sws_scale() workflow otherwise; a unit
test covers the new dynamic mode.

Verified locally: full test suite passes against system FFmpeg 9.0.1
(39+39+30) and a locally built FFmpeg 7.1 (38+37+30, new API compiled
out); zero errors when type-checking all targets against a
Windows-ABI-simulated binding (SwsFlags = c_int) and against the
FFmpeg 8.1 prebuilt binding for ffmpeg7_1/8/8_1.

+++ tests/misc/tutorial01.rs
@@ -80,7 +80,7 @@ fn _main(file: &CStr, out_dir: &str) -> Result<()> {
         decode_context.width,
         decode_context.height,
         ffi::AV_PIX_FMT_RGB24,
-        ffi::SWS_BILINEAR,
+        ffi::SWS_BILINEAR as u32,
         None,
         None,
         None,
